+++ src/parsers/mod.rs
@@ -4,6 +4,7 @@ pub mod dynamic_attr_records;
 pub mod general_xml;
 pub mod jproperties;
 pub mod magic;
+pub mod psm_tables;
 pub mod string_scan;
 pub mod xml_util;