+++ src/streams.rs
@@ -1,5 +1,6 @@
 pub mod cluster;
 pub mod dynamic_attrs;
 pub mod jsite;
+pub mod psm_tables;
 pub mod summary;
 pub mod tagged_text;